# Conflicts:
#	Source/Contrib/ActivityEditor/AEWizard/AEWizard.csproj
#	Source/Contrib/ActivityEditor/ActivityEditor/ActivityEditor.csproj
#	Source/Contrib/ActivityEditor/LibAE/LibAE.csproj
#	Source/Contrib/ContentManager/ContentManager.csproj
#	Source/Contrib/DataCollector/DataCollector.csproj
#	Source/Contrib/DataConverter/DataConverter.csproj
#	Source/Contrib/TrackViewer/Drawing/DrawTerrain.cs
#	Source/Contrib/TrackViewer/Drawing/ShadowDrawArea.cs
#	Source/Contrib/TrackViewer/TrackViewer.cs
#	Source/Contrib/TrackViewer/TrackViewer.csproj
#	Source/Menu/Menu.csproj
#	Source/ORTS.Common/ORTS.Common.csproj
#	Source/ORTS.Content/ORTS.Content.csproj
#	Source/ORTS.IO/ORTS.IO.csproj
#	Source/ORTS.Menu/ORTS.Menu.csproj
#	Source/ORTS.Parsers/ORTS.Parsers.csproj
#	Source/ORTS.Settings/ORTS.Settings.csproj
#	Source/ORTS.Updater/ORTS.Updater.csproj
#	Source/Orts.Formats.Msts/Orts.Formats.Msts.csproj
#	Source/Orts.Formats.OR/Orts.Formats.OR.csproj
#	Source/Orts.Parsers.Msts/Orts.Parsers.Msts.csproj
#	Source/Orts.Simulation/Common/Scripting/ScriptManager.cs
#	Source/Orts.Simulation/Orts.Simulation.csproj
#	Source/Orts.Simulation/Simulation/RollingStocks/MSTSDieselLocomotive.cs
#	Source/RunActivity/Content/SceneryShader.fx
#	Source/RunActivity/RunActivity.csproj
#	Source/RunActivity/Viewer3D/Precipitation.cs
#	Source/RunActivity/Viewer3D/RenderFrame.cs
#	Source/RunActivity/Viewer3D/RollingStock/MSTSDieselLocomotiveViewer.cs
#	Source/RunActivity/Viewer3D/RollingStock/MSTSLocomotiveViewer.cs
#	Source/RunActivity/Viewer3D/RollingStock/MSTSSteamLocomotiveViewer.cs
#	Source/RunActivity/Viewer3D/Transfers.cs
#	Source/RunActivity/Viewer3D/Viewer.cs
#	Source/RunActivity/Viewer3D/Weather.cs
#	Source/Tests/Tests.csproj
#	Source/Updater/Updater.csproj

+++ Source/Orts.Formats.Msts/TerrainFile.cs
@@ -185,7 +185,7 @@ namespace Orts.Formats.Msts
             A = block.ReadInt();
             B = block.ReadInt();
             C = block.ReadInt();
-            D = block.ReadInt();
+            D = (int)block.ReadFloat();
         }
     }
